+++ asafov.andrew/P3/matrixDynamicInitialization.cpp
@@ -8,7 +8,7 @@ void asafov::matrixDynamicInitialization(int* mtx, char* input, size_t& n, size_
   fin >> n;
   fin >> m;
   mtx = new int[n * m];
-  for (int i = 0; i < n * m; i++) {
+  for (size_t i = 0; i < n * m; i++) {
     fin >> mtx[i];
     if (fin.fail() || fin.eof()) {
       throw std::logic_error("err");
